+++ flask_restplus/api.py
@@ -3,6 +3,7 @@ from __future__ import unicode_literals
 
 import difflib
 import inspect
+from itertools import chain
 import logging
 import operator
 import re
@@ -200,8 +201,8 @@ class Api(object):
         app.handle_user_exception = partial(self.error_router, app.handle_user_exception)
 
         if len(self.resources) > 0:
-            for resource, urls, kwargs in self.resources:
-                self._register_view(app, resource, *urls, **kwargs)
+            for resource, namespace, urls, kwargs in self.resources:
+                self._register_view(app, resource, namespace, *urls, **kwargs)
 
         self._register_apidoc(app)
         self._validate = self._validate if self._validate is not None else app.config.get('RESTPLUS_VALIDATE', False)
@@ -238,6 +239,7 @@ class Api(object):
             self._register_view(
                 app_or_blueprint,
                 SwaggerView,
+                self.default_namespace,
                 '/swagger.json',
                 endpoint=endpoint,
                 resource_class_args=(self, )
@@ -258,12 +260,12 @@ class Api(object):
         self.endpoints.add(endpoint)
 
         if self.app is not None:
-            self._register_view(self.app, resource, *urls, **kwargs)
+            self._register_view(self.app, resource, namespace, *urls, **kwargs)
         else:
-            self.resources.append((resource, urls, kwargs))
+            self.resources.append((resource, namespace, urls, kwargs))
         return endpoint
 
-    def _register_view(self, app, resource, *urls, **kwargs):
+    def _register_view(self, app, resource, namespace, *urls, **kwargs):
         endpoint = kwargs.pop('endpoint', None) or camel_to_dash(resource.__name__)
         resource_class_args = kwargs.pop('resource_class_args', ())
         resource_class_kwargs = kwargs.pop('resource_class_kwargs', {})
@@ -272,17 +274,20 @@ class Api(object):
         if endpoint in getattr(app, 'view_functions', {}):
             previous_view_class = app.view_functions[endpoint].__dict__['view_class']
 
-            # if you override the endpoint with a different class, avoid the collision by raising an exception
+            # if you override the endpoint with a different class, avoid the
+            # collision by raising an exception
             if previous_view_class != resource:
-                msg = 'This endpoint (%s) is already set to the class %s.' % (endpoint, previous_view_class.__name__)
-                raise ValueError(msg)
+                msg = 'This endpoint (%s) is already set to the class %s.'
+                raise ValueError(msg % (endpoint, previous_view_class.__name__))
 
         resource.mediatypes = self.mediatypes_method()  # Hacky
         resource.endpoint = endpoint
+
         resource_func = self.output(resource.as_view(endpoint, self, *resource_class_args,
             **resource_class_kwargs))
 
-        for decorator in self.decorators:
+        # Apply Namespace and Api decorators to a resource
+        for decorator in chain(namespace.decorators, self.decorators):
             resource_func = decorator(resource_func)
 
         for url in urls: